Temporary commit, don't try to run mps tests with this one, they are extended... take more than one hour on my computer. Trying to figure out a BDCSVD & AVX2 bug... also fixed some compiler warnings.

+++ QCSim/QubitRegister.h
@@ -236,7 +236,7 @@ namespace QC {
 			std::vector<double> probabilities(registerStorage.size());
 
 			double accum = 0;
-			for (size_t i = 0; i < registerStorage.size(); ++i)
+			for (size_t i = 0; i < static_cast<size_t>(registerStorage.size()); ++i)
 			{
 				accum += std::norm(registerStorage[i]);
 				probabilities[i] = accum;
@@ -322,7 +322,7 @@ namespace QC {
 			std::vector<double> probabilities(registerStorage.size());
 
 			double accum = 0;
-			for (size_t i = 0; i < registerStorage.size(); ++i)
+			for (size_t i = 0; i < static_cast<size_t>(registerStorage.size()); ++i)
 			{
 				accum += std::norm(registerStorage[i]);
 				probabilities[i] = accum;
